feat(be): get values from on-chain proposal
